pyproject.toml
[project]
name = "pystuderxcom"
version = "3.2.0"
authors = [
  { name="Anko Hanse", email="anko_hanse@hotmail.com" },
]
description = "Python library to communicate with Studer-Innotec devices on the local network via Xcom protocol"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
  'httpx',
  'asyncudp',
  'pyserial',
  'pyserial-asyncio',
  'aiofiles',
  'orjson',
]

[project.optional-dependencies]
tests = [
  'pytest',
  'pytest-asyncio',
]
 
[project.urls]
Homepage = "https://github.com/ankohanse/pystuderxcom"
Issues = "https://github.com/ankohanse/pystuderxcom/issues"

[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "cython",
    "unasyncd",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pystuderxcom = ["*.json"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
